Move helm defun to try to fix init bug
Somehow, yet again, I triggered the init bug that complains about
helm-build-sync-source being an invalid function.  To make it weirder,
the bug went away when I deleted the .elc file that was built when I
upgraded the package, and it stayed away when I byte-compiled the file
manually.  I have no idea what's going on, and no idea how I manage to
keep retriggering this bug.